Calculadora

+++ Extacion de mio/clase12/OperacionesMatematicas.java
@@ -2,44 +2,44 @@ package clase12;
 
 public class OperacionesMatematicas {
 
-    //Atributos
+    // Atributos
     public int a;
     public int b;
- 
-    //Constructor de la clase
-    public OperacionesMatematicas (int a, int b){
+
+    // Constructor de la clase
+    public OperacionesMatematicas(int a, int b) {
         this.a = a;
         this.b = b;
     }
- 
-    //Metodo de sumar
-    public int sumar(int a, int b){
+
+    // Metodo de sumar
+    public int sumar(int a, int b) {
         return a + b;
     }
-   
-    //Metodo de restar
-    public int restar(int a, int b){
+
+    // Metodo de restar
+    public int restar(int a, int b) {
         return a - b;
     }
- 
-    //Metodo de multiplicar
-    public int multiplicar(int a, int b){
+
+    // Metodo de multiplicar
+    public int multiplicar(int a, int b) {
         return a * b;
     }
- 
-    //Metodo de dividir
-    public double dividir(int a, int b){
+
+    // Metodo de dividir
+    public double dividir(int a, int b) {
         double total;
-        if(b == 0){
+        if (b == 0) {
             total = 0.0;
-        }else{
-            total = (double) a / b;
+        } else {
+            total = (double) a / (double) b;
         }
         return total;
     }
- 
-    //Metodo de potencia
-    public int potencia (int a, int b){
+
+    // Metodo de potencia
+    public int potencia(int a, int b) {
         int p = 1;
         for (int i = 1; i <= b; i++) {
             p = p * a;
